# lib/openapi/generated.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "RealEST Property Marketplace API",
    "description": "Nigerian property marketplace API with ML validation, geospatial search, and intelligent matching",
    "version": "2.0.0",
    "contact": {
      "name": "RealEST Support",
      "email": "support@realest.ng"
    },
    "license": {
      "name": "Proprietary"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    },
    {
      "url": "https://realest.ng",
      "description": "Production server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Supabase JWT Auth token"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "const": false
          },
          "error": {
            "type": "string",
            "description": "Error message"
          },
          "code": {
            "type": "string",
            "description": "Error code"
          },
          "details": {
            "type": "object",
            "description": "Error details"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "success",
          "error",
          "timestamp"
        ]
      },
...
    }
  },
  "paths": {
...
    "/api/properties/{id}/media": {
      "get": {
        "method": "get",
        "summary": "Get property media files",
        "description": "Retrieve all media (images, videos) associated with a property. Published properties are visible to all users; draft properties only to the owner.",
        "tags": [
          "Properties - Media"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Property ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Media list retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "media": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "property_id": {
                            "type": "string",
                            "format": "uuid"
                          },
                          "media_type": {
                            "type": "string",
                            "enum": [
                              "image",
                              "video"
                            ]
                          },
                          "media_url": {
                            "type": "string",
                            "format": "url"
                          },
                          "is_featured": {
                            "type": "boolean"
                          },
                          "display_order": {
                            "type": "integer"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Property not found or not accessible",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "operationId": "getPropertiesIdMedia"
      },
      "post": {
        "method": "post",
        "summary": "Upload media for property",
        "description": "Upload a new media file (image or video) to a property. Only property owner can upload. Media can only be added to draft properties.",
        "tags": [
          "Properties - Media"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Property ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "media_type",
                  "media_url"
                ],
                "properties": {
                  "media_type": {
                    "type": "string",
                    "enum": [
                      "image",
                      "video"
                    ],
                    "description": "Type of media"
                  },
                  "media_url": {
                    "type": "string",
                    "format": "url",
                    "description": "URL to media file"
                  },
                  "is_featured": {
                    "type": "boolean",
                    "description": "Mark as primary/featured image"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Media uploaded successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "media": {
                      "type": "object"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request or property is published"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Property not found or access denied"
          }
        },
        "operationId": "postPropertiesIdMedia"
      }
    },
...
    "/api/referral/resolve": {
      "get": {
        "method": "get",
        "summary": "Resolve referral code",
        "description": "Resolve a referral code to inviter details and milestone context.",
        "tags": [
          "Referral"
        ],
        "parameters": [
          {
            "name": "code",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Referral code resolved"
          },
          "400": {
            "description": "Referral code is required"
          },
          "404": {
            "description": "Referral code not found"
          }
        },
        "operationId": "getReferralResolve"
      }
    },
...
    "/api/upload/signed-url": {
      "post": {
        "method": "post",
        "summary": "Generate signed upload URL",
        "description": "Create a signed URL for direct file upload to storage.",
        "tags": [
          "Upload"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Signed URL generated"
          },
          "400": {
            "description": "Invalid file data"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "operationId": "postUploadSignedUrl"
      }
    },
...
  },
...
}